generate toc at build time
