+++ src/components/RightSidebar/RightSidebar.astro
@@ -3,11 +3,13 @@ import TableOfContents from './TableOfContents';
 import ContributeMenu from './ContributeMenu.astro';
 import CommunityMenu from './CommunityMenu.astro';
 import { useTranslations } from '../../i18n/util';
+import generateToc from '~/util/generateToc';
 
 const t = useTranslations(Astro);
 
 const { content, githubEditUrl } = Astro.props;
 const headings = content.astro?.headings;
+const overview = t('rightSidebar.overview');
 ---
 
 <nav aria-label={t('rightSidebar.a11yTitle')}>
@@ -15,8 +17,8 @@ const headings = content.astro?.headings;
 		headings && (
 			<TableOfContents
 				client:media="(min-width: 50em)"
-				headings={headings}
-				labels={{ onThisPage: t('rightSidebar.onThisPage'), overview: t('rightSidebar.overview') }}
+				toc={generateToc(headings, overview)}
+				labels={{ onThisPage: t('rightSidebar.onThisPage'), overview: overview }}
 				isMobile={false}
 			/>
 		)
