+++ src/layouts/MainLayout.astro
@@ -1,4 +1,5 @@
 ---
+import generateToc from '~/util/generateToc';
 import PageContent from '../components/PageContent/PageContent.astro';
 import RightSidebar from '../components/RightSidebar/RightSidebar.astro';
 import TableOfContents from '../components/RightSidebar/TableOfContents';
@@ -12,6 +13,7 @@ const headings = content.astro?.headings;
 const t = useTranslations(Astro);
 const { previous, next } = await getNavLinks(Astro);
 const githubEditUrl = getGithubEditUrl(Astro);
+const overview = t('rightSidebar.overview')
 ---
 
 <BaseLayout {...Astro.props}>
@@ -23,10 +25,10 @@ const githubEditUrl = getGithubEditUrl(Astro);
 					<nav>
 						<TableOfContents
 							client:media="(max-width: 72em)"
-							headings={headings}
+							toc={generateToc(headings, overview)}
 							labels={{
 								onThisPage: t('rightSidebar.onThisPage'),
-								overview: t('rightSidebar.overview'),
+								overview: overview,
 							}}
 							isMobile={true}
 						/>
